+++ lib/src/solid/utils/init_helper.dart
